requirements.txt
asyncio==4.0.0
websockets==15.0.1
pytz==2025.2
fastapi==0.121.2
uvicorn==0.38.0
twilio==9.8.6
google-auth-oauthlib==1.2.3
google-cloud-aiplatform==1.127.0
google-cloud-speech==2.34.0
google-cloud-texttospeech==2.33.0
google-cloud-secret-manager==2.25.0
requests==2.32.5
python-dotenv==1.2.1
google-api-python-client==2.187.0
numpy==2.3.4
sounddevice==0.5.3
webrtcvad==2.0.10
google-cloud-storage==3.5.0
python-multipart
google-genai==1.50.1
pyaudio

websockets

fastapi 
uvicorn 
python-dotenv 
websockets 
twilio 
google-auth-oauthlib 
google-cloud-secret-manager 
requests
